+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "cl-wordle"
 authors = ["Conrad Ludgate <conradludgate@gmail.com>"]
-version = "0.1.10"
+version = "0.2.0"
 edition = "2021"
 license = "MIT"
 description = "Wordle in your terminal"
@@ -10,8 +10,9 @@ repository = "https://github.com/conradludgate/wordle"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+eyre = { version = "0.6" }
 time = { version = "0.3", features = ["local-offset", "macros", "parsing"], optional = true }
-eyre = { version = "0.6", optional = true }
+
 owo-colors = { version = "3.2", optional = true }
 crossterm = { version = "0.23", optional = true }
 clap = { version = "3.0", features = ["derive"], optional = true }
@@ -21,9 +22,10 @@ rand = { version = "0.8", optional = true }
 test-case = "1.2.2"
 
 [features]
-default = ["tui", "time"]
-tui = ["time", "eyre", "owo-colors", "crossterm", "clap", "rand"]
+default = ["tui", "rand"]
+tui = ["owo-colors", "crossterm", "cli"]
+cli = ["time", "clap"]
 
 [[bin]]
 name = "wordle"
-required-features = ["tui"]
+required-features = ["cli"]
