+++ src/bin/wordle/main.rs
@@ -1,16 +1,13 @@
+use cl_wordle::game::Game;
 use clap::Parser;
 
 mod args;
 mod controller;
-mod game;
-mod guess;
-mod keyboard;
-mod letters;
-mod terminal;
 
 use args::{App, GameMode};
-use controller::GameController;
-use game::Game;
+use controller::cli;
+
+#[cfg(feature = "rand")]
 use rand::Rng;
 
 fn main() -> eyre::Result<()> {
@@ -19,12 +16,21 @@ fn main() -> eyre::Result<()> {
         None => Game::new()?,
         Some(GameMode::Custom(custom)) => Game::custom(custom.word)?,
         Some(GameMode::Day(day)) => Game::from_day(day.day)?,
+        #[cfg(feature = "rand")]
         Some(GameMode::Random) => Game::from_day(rand::thread_rng().gen())?,
         Some(GameMode::Date(date)) => Game::from_date(date.date)?,
     };
 
-    let controller = GameController::new(game)?;
-    if let Some(share) = controller.run()? {
+    #[cfg(feature = "tui")]
+    let output = if app.no_tui {
+        Some(cli::Controller::new(game).run()?)
+    } else {
+        controller::tui::Controller::new(game)?.run()?
+    };
+    #[cfg(not(feature = "tui"))]
+    let output = Some(cli::Controller::new(game).run()?);
+
+    if let Some(share) = output {
         println!("{}", share);
     }
